# internal/api/gen/api.gen.go
//go:build go1.22

// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
)

// Defines values for MetaService.
const (
	Opsgrid MetaService = "opsgrid"
)

// Valid indicates whether the value is a known member of the MetaService enum.
func (e MetaService) Valid() bool {
	switch e {
	case Opsgrid:
		return true
	default:
		return false
	}
}

// Defines values for ProblemCode.
const (
	AUTHENTICATIONREQUIRED ProblemCode = "AUTHENTICATION_REQUIRED"
	DEPENDENCYUNAVAILABLE  ProblemCode = "DEPENDENCY_UNAVAILABLE"
	FORBIDDEN              ProblemCode = "FORBIDDEN"
	IDEMPOTENCYCONFLICT    ProblemCode = "IDEMPOTENCY_CONFLICT"
	IDEMPOTENCYKEYREQUIRED ProblemCode = "IDEMPOTENCY_KEY_REQUIRED"
	INTERNALERROR          ProblemCode = "INTERNAL_ERROR"
	INVALIDTRANSITION      ProblemCode = "INVALID_TRANSITION"
	PRECONDITIONREQUIRED   ProblemCode = "PRECONDITION_REQUIRED"
	RATELIMITED            ProblemCode = "RATE_LIMITED"
	REQUESTTOOLARGE        ProblemCode = "REQUEST_TOO_LARGE"
	RESOURCENOTFOUND       ProblemCode = "RESOURCE_NOT_FOUND"
	SCHEDULECONFLICT       ProblemCode = "SCHEDULE_CONFLICT"
	TENANTNOTFOUND         ProblemCode = "TENANT_NOT_FOUND"
	UNSUPPORTEDMEDIATYPE   ProblemCode = "UNSUPPORTED_MEDIA_TYPE"
	VALIDATIONFAILED       ProblemCode = "VALIDATION_FAILED"
	VERSIONCONFLICT        ProblemCode = "VERSION_CONFLICT"
)

// Valid indicates whether the value is a known member of the ProblemCode enum.
func (e ProblemCode) Valid() bool {
	switch e {
	case AUTHENTICATIONREQUIRED:
		return true
	case DEPENDENCYUNAVAILABLE:
		return true
	case FORBIDDEN:
		return true
	case IDEMPOTENCYCONFLICT:
		return true
	case IDEMPOTENCYKEYREQUIRED:
		return true
	case INTERNALERROR:
		return true
	case INVALIDTRANSITION:
		return true
	case PRECONDITIONREQUIRED:
		return true
	case RATELIMITED:
		return true
	case REQUESTTOOLARGE:
		return true
	case RESOURCENOTFOUND:
		return true
	case SCHEDULECONFLICT:
		return true
	case TENANTNOTFOUND:
		return true
	case UNSUPPORTEDMEDIATYPE:
		return true
	case VALIDATIONFAILED:
		return true
	case VERSIONCONFLICT:
		return true
	default:
		return false
	}
}

// FieldError defines model for FieldError.
type FieldError struct {
	Field   string `json:"field"`
	Message string `json:"message"`
}

// Meta defines model for Meta.
type Meta struct {
	// Commit Abbreviated git commit of the build.
	Commit  string      `json:"commit"`
	Service MetaService `json:"service"`

	// Version Examples: v1.0.0, dev
	Version string `json:"version"`
}

// MetaService defines model for Meta.Service.
type MetaService string

// Problem RFC 9457 problem details with OpsGrid extension members.
type Problem struct {
	// Code Stable, machine-readable error identifier (never renamed).
	Code      ProblemCode             `json:"code"`
	Detail    *string                 `json:"detail,omitempty"`
	Details   *map[string]interface{} `json:"details,omitempty"`
	Errors    *[]FieldError           `json:"errors,omitempty"`
	Instance  *string                 `json:"instance,omitempty"`
	RequestId *string                 `json:"requestId,omitempty"`
	Status    int                     `json:"status"`
	Title     string                  `json:"title"`
	TraceId   *string                 `json:"traceId,omitempty"`
	Type      string                  `json:"type"`
}

// ProblemCode Stable, machine-readable error identifier (never renamed).
type ProblemCode string

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetMeta Service build information
	// (GET /v1/meta)
	GetMeta(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetMeta operation middleware
func (siw *ServerInterfaceWrapper) GetMeta(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetMeta(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/v1/meta", wrapper.GetMeta)

	return m
}

type ProblemApplicationProblemPlusJSONResponse Problem

type GetMetaRequestObject struct {
}

type GetMetaResponseObject interface {
	VisitGetMetaResponse(w http.ResponseWriter) error
}

type GetMeta200JSONResponse Meta

func (response GetMeta200JSONResponse) VisitGetMetaResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetMetadefaultApplicationProblemPlusJSONResponse struct {
	Body       Problem
	StatusCode int
}

func (response GetMetadefaultApplicationProblemPlusJSONResponse) VisitGetMetaResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// GetMeta Service build information
	// (GET /v1/meta)
	GetMeta(ctx context.Context, request GetMetaRequestObject) (GetMetaResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetMeta operation middleware
func (sh *strictHandler) GetMeta(w http.ResponseWriter, r *http.Request) {
	var request GetMetaRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetMeta(ctx, request.(GetMetaRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetMeta")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetMetaResponseObject); ok {
		if err := validResponse.VisitGetMetaResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}