# packages/api-client/src/gen/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

export type Meta = {
    service: 'opsgrid';
    version: string;
    /**
     * Abbreviated git commit of the build.
     */
    commit: string;
};

/**
 * Stable, machine-readable error identifier (never renamed).
 */
export type ProblemCode = 'AUTHENTICATION_REQUIRED' | 'FORBIDDEN' | 'TENANT_NOT_FOUND' | 'RESOURCE_NOT_FOUND' | 'VALIDATION_FAILED' | 'PRECONDITION_REQUIRED' | 'VERSION_CONFLICT' | 'SCHEDULE_CONFLICT' | 'INVALID_TRANSITION' | 'IDEMPOTENCY_CONFLICT' | 'IDEMPOTENCY_KEY_REQUIRED' | 'REQUEST_TOO_LARGE' | 'UNSUPPORTED_MEDIA_TYPE' | 'RATE_LIMITED' | 'DEPENDENCY_UNAVAILABLE' | 'INTERNAL_ERROR';

export type FieldError = {
    field: string;
    message: string;
};

/**
 * RFC 9457 problem details with OpsGrid extension members.
 */
export type Problem = {
    type: string;
    title: string;
    status: number;
    detail?: string;
    instance?: string;
    code: ProblemCode;
    requestId?: string;
    traceId?: string;
    details?: {
        [key: string]: unknown;
    };
    errors?: Array<FieldError>;
};

export type GetMetaData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/v1/meta';
};

export type GetMetaErrors = {
    /**
     * An RFC 9457 problem document.
     */
    default: Problem;
};

export type GetMetaError = GetMetaErrors[keyof GetMetaErrors];

export type GetMetaResponses = {
    /**
     * Build information.
     */
    200: Meta;
};

export type GetMetaResponse = GetMetaResponses[keyof GetMetaResponses];